Added the Dummy servo file

+++ servocontroller/src/Servo/USB.cpp
@@ -50,7 +50,7 @@ short int Servo_USB::getError()
         throw Exception_Servo();
     
     //TODO This needs to be fixed, its wrong at the moment
-    return (int)sqrt(response[0] + 256*response[1]);
+    return (short int)sqrt(response[0] + 256*response[1]);
 }
 
 /**
@@ -71,7 +71,7 @@ short int Servo_USB::getError()
  *
  * @return int - The value of the channel 
  */
-short int Servo_USB::getTarget(const unsigned char channel)
+unsigned short int Servo_USB::getTarget(const unsigned char channel)
 {
     unsigned char command[2];
     command[0] = 0x90;
@@ -102,7 +102,7 @@ short int Servo_USB::getTarget(const unsigned char channel)
  *
  * @throw Exception_Servo - If writing to the board fails
  */
-void Servo_USB::setTarget(const unsigned char channel, short int target)
+void Servo_USB::setTarget(const unsigned char channel, unsigned short int target)
 {
     calculateTarget(target);
